+++ api/src/infrastructure/film_record/__init__.py
@@ -1,0 +1,1 @@
+from .film_record_repository import ImplFilmRecordRepository
