+++ api/src/infrastructure/film_record/genre/__init__.py
@@ -1,0 +1,1 @@
+from .film_genre_repository import ImplFilmGenreRepository
