+++ api/src/infrastructure/film_record/genre/inmemory/__init__.py
@@ -1,0 +1,1 @@
+from .inmemory_film_genre_repository import ImplInmemoryFilmGenreRepository
