+++ api/src/infrastructure/film_record/inmemory/__init__.py
@@ -1,0 +1,1 @@
+from .inmemory_film_record_repository import ImplInmemoryFilmRecordRepository
