+++ api/src/infrastructure/film_record/inmemory/inmemory_film_record_repository.py
@@ -13,7 +13,7 @@ from src.domain.film_record.film import FilmEntity, TmdbIdObject
 from src.domain.film_record.film.genre import FilmGenreEnum
 from src.domain.film_record.film.poster import FilmPosterObject
 from src.domain.film_record.film.series import FilmSeriesObject
-from src.domain.film_record.watch_medium.watch_medium_enum import WatchMediumEnum
+from src.domain.film_record.watch_medium import WatchMediumEnum
 
 logger = getLogger(__name__)
