+++ api/src/infrastructure/film_record/series/__init__.py
@@ -1,0 +1,1 @@
+from .film_series_repository import ImplFilmSeriesRepository
