+++ api/src/infrastructure/film_record/series/film_series_repository.py
@@ -1,6 +1,10 @@
+# 標準ライブラリ
 import os
 
+# 外部ライブラリ
 from notion_client import Client
+
+# 独自ライブラリ
 from src.domain.film_record.film.series.film_series_repository import IFilmSeriesRepository
