+++ api/src/infrastructure/film_record/series/inmemory/__init__.py
@@ -1,0 +1,1 @@
+from .inmemory_film_series_repository import ImplInmemoryFilmSeriesRepository
