+++ api/src/infrastructure/film_record/watch_history/__init__.py
@@ -1,0 +1,1 @@
+from .watch_history_repository import ImplFilmWatchHistoryRepository
