+++ api/src/infrastructure/film_record/watch_history/inmemory/__init__.py
@@ -1,0 +1,1 @@
+from .inmemory_watch_history_repository import ImplInmemoryFilmWatchHistoryRepository
