+++ api/src/infrastructure/film_record/watch_medium/__init__.py
@@ -1,0 +1,1 @@
+from .watch_medium_repository import ImplWatchMediumRepository
