+++ api/src/infrastructure/film_record/watch_medium/inmemory/__init__.py
@@ -1,0 +1,1 @@
+from .inmemory_watch_medium_repository import ImplInmemoryWatchMediumRepository
