+++ api/src/infrastructure/film_record/watch_medium/watch_medium_repository.py
@@ -1,7 +1,11 @@
+# 標準ライブラリ
 import os
 
+# 外部ライブラリ
 from notion_client import Client
-from src.domain.film_record.watch_medium.watch_medium_repository import IWatchMediumRepository
+
+# 独自ライブラリ
+from src.domain.film_record.watch_medium import IWatchMediumRepository
 
 
 class ImplWatchMediumRepository(IWatchMediumRepository):
